+++ Backend/Application/Dashboard/Dtos/DashboardSale.cs
@@ -3,5 +3,5 @@ namespace Application.Dashboard.Dtos;
 public class DashboardSale
 {
     public DateTime Date { get; set; }
-    public double Revenue { get; set; }
+    public decimal Revenue { get; set; }
 }
